Randomness and Inefficiencies Detector/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "v75-randomness-detector"
version = "0.1.0"
description = "Standalone external CLI for V75 randomness and inefficiency analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.1.0",
  "numpy>=1.24.0",
  "scipy>=1.11.0",
  "statsmodels>=0.14.0",
  "arch>=7.0.0",
  "matplotlib>=3.8.0",
  "PyYAML>=6.0.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
]

[project.scripts]
rid = "rid.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
